requirements.txt
requests>=2.25.1
Pillow>=9.0.0